Move omop models
