+++ app/api/config/settings.py
@@ -59,8 +59,8 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "extra_views",
     "api",
+    "shared.data",
     "shared.mapping",
-    "data",
     "rest_framework",
     "django_filters",
     "rest_framework.authtoken",
