+++ app/shared/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "shared"
-version = "0.3.0"
+version = "0.3.1"
 description = "Carrot shared library"
 authors = ["Sam Cox <sam.cox@nottingham.ac.uk>", "Philip Quinlan <philip.quinlan@nottingham.ac.uk>"]
 license = "MIT"
