+++ app/shared/shared/data/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DataConfig(AppConfig):
+    name = "shared.data"
+    label = "data"
