--- app/shared/shared/data/omop/__init__.py
@@ -1,8 +1,0 @@
-from .concept import Concept
-from .concept_ancestor import ConceptAncestor
-from .concept_class import ConceptClass
-from .concept_relationship import ConceptRelationship
-from .concept_synonym import ConceptSynonym
-from .domain import Domain
-from .drug_strength import DrugStrength
-from .vocabulary import Vocabulary
