+++ app/shared/shared/services/rules_export.py
@@ -7,7 +7,7 @@ from django.contrib.contenttypes.models import ContentType
 from django.db.models import Q
 from django.db.models.query import QuerySet
 from graphviz import Digraph
-from shared.data.omop import Concept, ConceptAncestor
+from shared.data.models import Concept, ConceptAncestor
 from shared.mapping.models import (
     MappingRule,
     OmopField,
